# api-specs.yaml
openapi: 3.0.3
info:
  title: Shadow Ledger System API
  description: |
    Banking backend system for processing financial events, maintaining shadow ledger,
    and detecting drift between CBS and shadow balances.
  version: 1.0.0
  contact:
    name: Shadow Ledger Team
servers:
  - url: http://localhost:8080
    description: API Gateway (Local Development)
...
paths:
...
  /accounts/{accountId}/shadow-balance:
    get:
      tags:
        - Shadow Ledger
      summary: Get shadow balance
      description: Retrieve the computed shadow balance for an account
      operationId: getShadowBalance
      security:
        - BearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
          description: Account identifier
          example: "A10"
        - $ref: '#/components/parameters/TraceId'
      responses:
        '200':
          description: Shadow balance retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShadowBalance'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/token endpoint
  parameters:
    TraceId:
      name: X-Trace-Id
      in: header
      required: false
      schema:
        type: string
      description: Distributed tracing ID
      example: "7f3a2b1c-4d5e-6f7g-8h9i-0j1k2l3m4n5o"
  schemas:
...
    ShadowBalance:
      type: object
      properties:
        accountId:
          type: string
          example: "A10"
        balance:
          type: number
          format: decimal
          example: 750.00
        lastEvent:
          type: string
          example: "E005"
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "invalid_credentials"
  responses:
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Unauthorized"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Forbidden - insufficient role"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
security:
  - BearerAuth: []